move time integrator out
