+++ src/CMakeLists.txt
@@ -201,8 +201,9 @@ add_library(parthenon
   tasks/task_list.hpp
   tasks/task_types.hpp
 
-  time_integration/staged_integrator.hpp
+  time_integration/low_storage_integrator.cpp
   time_integration/staged_integrator.cpp
+  time_integration/staged_integrator.hpp
 
   utils/alias_method.cpp
   utils/alias_method.hpp
